+++ tests/jax/multi_chip/bounties/falcon3_7b/test/generate_hf.py
@@ -1,5 +1,6 @@
 import torch
-from transformers import AutoTokenizer, AutoModelForCausalLM, AutoConfig
+from transformers import AutoConfig
+from utils.torch_utils import *
 
 MODEL_NAME = "tiiuae/Falcon3-7B-Instruct"
 EXAMPLE_PROMPT = """
@@ -8,46 +9,14 @@ She sells the remainder at the farmers' market daily for $2 per fresh duck egg.
 How much in dollars does she make every day at the farmers' market?\n
 A: """
 
-def init_torch_model(model_name: str, config):
-    """
-    Initialize the PyTorch model with the given configuration.
-    """
-    torch_model = AutoModelForCausalLM.from_pretrained(
-        model_name,
-        config=config,
-        device_map="auto",
-        torch_dtype=torch.float32,
-    )
-    return torch_model
-
-def prepare_torch_input(model_name, prompt):
-    """
-    Prepare input for the PyTorch model.
-    """
-    tokenizer = AutoTokenizer.from_pretrained(model_name)
-    inputs = tokenizer(prompt, return_tensors="pt")
-    
-    return tokenizer, inputs.input_ids, inputs.attention_mask
-
-def run_torch_model(torch_model, input_ids, attention_mask):
-    """
-    Run the PyTorch model with the given input IDs and attention mask.
-    """
-    print("🏢 Generating HF Model output...")
-    outputs = torch_model.generate(
-        input_ids=input_ids,
-        attention_mask=attention_mask,
-    )
-    return outputs
-
-def run_test(model_name: str, prompt: str):
+def main(model_name: str, prompt: str):
     """
     Run the test comparing Hugging Face and Flax models.
     """
     print("🪄  Initializing models...")
     config = AutoConfig.from_pretrained(
         model_name,
-        # num_hidden_layers=4, # for easeier testing
+        num_hidden_layers=2, # for easeier testing
         torch_dtype=torch.float32,
     )
     tokenizer, input_ids, attention_mask = prepare_torch_input(model_name, prompt)
@@ -59,7 +28,7 @@ def run_test(model_name: str, prompt: str):
     print("🈵 Decoded output:", torch_result[0])
 
 if __name__ == "__main__":
-    torch_output = run_test(
+    torch_output = main(
         model_name=MODEL_NAME,
         prompt=EXAMPLE_PROMPT
     )
